refs #25 ユースケースの改行を整える

+++ src/main/java/naitokikaku/sscoordinator/application/usecase/account/changeemail/ChangeEmailAddress.java
@@ -4,6 +4,7 @@ import naitokikaku.sscoordinator.application.usecase.account.changeemail.complet
 import naitokikaku.sscoordinator.domain.model.account.Account;
 import naitokikaku.sscoordinator.domain.model.account.AccountRepository;
 import naitokikaku.sscoordinator.domain.model.account.EmailAddress;
+import naitokikaku.sscoordinator.domain.model.account.policy.AccountPolicy;
 import naitokikaku.sscoordinator.domain.model.account.revision.AccountRevision;
 import naitokikaku.sscoordinator.domain.model.account.snapshot.AccountSnapshot;
 import org.springframework.context.ApplicationEventPublisher;
@@ -15,13 +16,18 @@ import javax.annotation.Resource;
 @Component
 public class ChangeEmailAddress {
     @Resource
+    AccountPolicy accountPolicy;
+    @Resource
     AccountRepository accountRepository;
     @Resource
     ApplicationEventPublisher publisher;
 
     @Transactional
     public void execute(AccountSnapshot snapshot, EmailAddress emailAddress) {
+        Account draft = snapshot.account().replace(emailAddress);
+        if (!accountPolicy.ok(draft)) throw new IllegalArgumentException();
         if (emailAddress.same(snapshot.emailAddress())) return;
+
         accountRepository.deleteActive(snapshot.emailAddress());
         accountRepository.storeActive(emailAddress);
         AccountRevision updatedRevision = accountRepository.update(emailAddress);
